calculator_server/requirements.txt
grpcio-tools==1.26.0